requirements.txt
praw~=7.8.0
python-dotenv~=1.0.1
streamlit~=1.36.0